requirements.txt
pandas>=1.3.0
rdkit-pypi>=2022.9.1